Added geo view in results pane.
